clean-up & renaming
