+++ timesketch/lib/analyzers/feature_extraction_plugin_test.py
@@ -22,8 +22,8 @@ import yaml
 import mock
 
 from timesketch.lib import emojis
-from timesketch.lib.analyzers.feature import FeatureSketchPlugin
-from timesketch.lib.analyzers.feature_plugins import regex_features
+from timesketch.lib.analyzers.feature_extraction_plugin import FeatureSketchPlugin
+from timesketch.lib.analyzers.feature_extraction_plugins import regex_features
 from timesketch.lib.analyzers.sequence_sessionizer_test import _create_eventObj
 from timesketch.lib.testlib import BaseTest
 from timesketch.lib.testlib import MockDataStore
