+++ timesketch/lib/analyzers/feature_extraction_plugins/__init__.py
@@ -13,5 +13,5 @@
 # limitations under the License.
 """Imports for the feature extraction plugins."""
 
-from timesketch.lib.analyzers.feature_plugins import winevt_features
-from timesketch.lib.analyzers.feature_plugins import regex_features
+from timesketch.lib.analyzers.feature_extraction_plugins import winevt_features
+from timesketch.lib.analyzers.feature_extraction_plugins import regex_features
