+++ timesketch/lib/analyzers/feature_extraction_plugins/manager.py
