+++ timesketch/lib/analyzers/feature_extraction_plugins/winevt_features.py
@@ -17,8 +17,8 @@ import logging
 from typing import List, Dict
 
 from timesketch.lib.analyzers import interface as base_interface
-from timesketch.lib.analyzers.feature_plugins import interface
-from timesketch.lib.analyzers.feature_plugins import manager
+from timesketch.lib.analyzers.feature_extraction_plugins import interface
+from timesketch.lib.analyzers.feature_extraction_plugins import manager
 
 logger = logging.getLogger("timesketch.analyzers.winevt_features")
